requirements.txt
imageio==2.36.0
nd2reader==3.3.0
numpy==2.1.2
packaging==24.1
pillow==11.0.0
PIMS==0.7
six==1.16.0
slicerator==1.1.0
tifffile==2024.9.20
xmltodict==0.14.1
